refactor(inscricao): adiciona gerenciamento dos estágios por roteamento
