+++ src/app/register/RegisterForm/index.jsx
@@ -1,125 +1,35 @@
 'use client';
 
 import styles from '@/app/register/RegisterForm/RegisterForm.module.css';
-import { useMemo, useState } from 'react';
-import AccountStage from './AccountStage';
-import AdditionalInfo from './AdditionalInfo';
-import Button from '@/components/Button';
 import { useLanguage } from '@/contexts/LanguageContext';
-import { ChevronLeft, ChevronRight } from 'lucide-react';
+
 import Image from 'next/image';
-import { useFormStage } from '@/hooks/useFormStage';
-import { createAdditionalInfoSchema } from '@/schemas/register.schema';
-import { createAccountSchema } from '@/schemas/register.schema';
-import { z } from 'zod';
-import { registerParticipant } from '@/lib/api';
+
 import { Loading } from '@/components/Loading';
 import { Error } from '@/components/Error';
+import { useRegister } from '@/contexts/RegisterContext';
+import { useRouter } from 'next/navigation';
 
-const schemas = [
-  {
-    create: createAccountSchema,
-    defaultValues: {
-      firstName: '',
-      lastName: '',
-      email: '',
-      password: '',
-      confirmPassword: '',
-    },
-  },
-  {
-    create: createAdditionalInfoSchema,
-    defaultValues: {
-      institution: '',
-      country: '',
-      state: '',
-      city: '',
-      gender: 'not declared',
-      modality: '',
-    },
-  },
-];
+const stageRoutes = ['account-info', 'additional-info', 'confirmation'];
 
-export function RegisterForm() {
-  const { texts, language } = useLanguage();
+export function RegisterForm({ children }) {
+  const { texts } = useLanguage();
+  const {
+    isLoading,
+    errorMessage,
+    resetErrorMessage,
+    currentStage,
+    goToStage,
+  } = useRegister();
 
-  const [isLoading, setIsLoading] = useState(false);
-  const [errorMessage, setErrorMessage] = useState(null);
-  const [currentStage, setCurrentStage] = useState(0);
-  const [formData, setFormData] = useState({});
+  const router = useRouter();
 
   const stages = texts.register.steps;
 
-  const handleStageValid = (data) => {
-    console.log(data);
-    setFormData((prev) => ({ ...prev, ...data }));
-  };
+  const handleStageClick = async (index) => {
+    const success = await goToStage(index);
 
-  const { create: createSchema, defaultValues } = useMemo(
-    () =>
-      schemas[currentStage] || {
-        create: () => z.object({}),
-        defaultValues: {},
-      },
-    [currentStage, language]
-  );
-
-  const schema = createSchema ? createSchema(language) : z.object({});
-
-  const formFunctions = useFormStage(schema, defaultValues, handleStageValid);
-
-  const handleValidateStage = async () => {
-    if (formFunctions.validate) {
-      const isValid = await formFunctions.validate();
-
-      if (!isValid) return false;
-    }
-
-    return true;
-  };
-
-  const handleChangeToStage = async (stage) => {
-    if (stage < 0 || stage > stages.length - 1) return;
-    if (stage === currentStage) return;
-
-    if (stage === stages.length - 1) {
-      const isValid = await handleValidateStage();
-
-      if (!isValid) return;
-
-      const currentFormData = formFunctions.getValues();
-      const allData = { ...formData, ...currentFormData };
-
-      setFormData(allData);
-
-      const { success, error } = await handleSubmit(allData);
-
-      if (success) {
-        setCurrentStage(stage);
-        setFormData({});
-        setErrorMessage(null);
-      }
-
-      setErrorMessage(error.message);
-
-      return;
-    }
-
-    if (stage > currentStage) {
-      const isValid = await handleValidateStage();
-
-      if (!isValid) return;
-    }
-
-    setCurrentStage(stage);
-  };
-
-  const handleSubmit = async (data) => {
-    setIsLoading(true);
-    const response = await registerParticipant({ language, ...data });
-    setIsLoading(false);
-
-    return response;
+    if (success) router.push(`/register/${stageRoutes[index]}`);
   };
 
   return (
@@ -131,7 +41,7 @@ export function RegisterForm() {
       )}
       {errorMessage && (
         <div className={styles.errorMessage}>
-          <Error message={errorMessage} onClick={() => setErrorMessage(null)} />
+          <Error message={errorMessage} onClick={resetErrorMessage} />
         </div>
       )}
       <div className={styles.images}>
@@ -156,7 +66,7 @@ export function RegisterForm() {
               <li
                 key={stage}
                 className={`${styles.navItem} ${index === currentStage ? styles.active : ''}`}
-                onClick={() => handleChangeToStage(index)}
+                onClick={() => handleStageClick(index)}
               >
                 {index === currentStage ? (
                   <Image
@@ -179,30 +89,7 @@ export function RegisterForm() {
           </ul>
         </nav>
       </header>
-      <div className={styles.form}>
-        {currentStage === 0 && (
-          <AccountStage styles={styles} formFunctions={formFunctions} />
-        )}
-        {currentStage === 1 && (
-          <AdditionalInfo styles={styles} formFunctions={formFunctions} />
-        )}
-        <div className={styles.buttons}>
-          <Button
-            variant='ghost'
-            onClick={() => handleChangeToStage(currentStage - 1)}
-          >
-            <ChevronLeft size={24} />
-            <span>{texts.register.buttons.back}</span>
-          </Button>
-          <Button
-            variant='filled'
-            onClick={() => handleChangeToStage(currentStage + 1)}
-          >
-            <span>{texts.register.buttons.continue}</span>
-            <ChevronRight size={24} />
-          </Button>
-        </div>
-      </div>
+      <div className={styles.form}>{children}</div>
     </section>
   );
 }
